+++ packages/gdsc-dju-blog/package.json
@@ -36,7 +36,6 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "^4.0.3",
     "react-syntax-highlighter": "^15.5.0",
-    "react-use": "^17.3.2",
     "recoil": "^0.7.4",
     "recoil-persist": "^4.0.0",
     "styled-components": "^5.3.3"
