+++ packages/gdsc-dju-blog/src/api/hooks/usePostBookMark.ts
@@ -1,0 +1,13 @@
+import { useMutation } from 'react-query';
+import PostService from '../PostService';
+
+export const usePostBookMark = () => {
+  const {
+    data: bookMarkResponse,
+    mutate,
+    isLoading,
+    isSuccess,
+  } = useMutation(PostService.updateMyScrapData);
+
+  return { bookMarkResponse, mutate, isLoading, isSuccess };
+};
